+++ CHANGELOG.md
@@ -44,6 +44,8 @@ Main (unreleased)
   - `otelcol.exporter.loki` forwards OTLP-formatted data to compatible `loki`
     receivers. (@tpaschalis)
   - `loki.source.gelf` listens for Graylog logs. (@mattdurham)
+  - `loki.source.journal` read messages from systemd journal. (@mattdurham)
+